Create Shortest path in undirected graph having unit distance
Approach:
All the edges are undirected with a unit distance. so the approach is doing a simple bfs using queue. here we need a distnce array of size of number of edges in which initially we store maximum distance. then after starting from the source node and pushing it into queue with its distance it is carrying. further we will mark dist of src node as 0 bcz it do not carry any distance. then one by one we weill pop out node from the queue with its distance and iterate over its neighbour nodes and update distance id distance [it] + 1 < distance [it] until the queue gets empty. and at last we will get a distance array with the shortest path.

Time Complexity: O(M) { for creating the adjacency list from given list ‘edges’} + O(N + 2M) { for the BFS Algorithm} + O(N) { for adding the final values of the shortest path in the resultant array} ~ O(N+2M). 

Where N= number of vertices and M= number of edges.

Space Complexity:  O( N) {for the stack storing the BFS} + O(N) {for the resultant array} + O(N) {for the dist array storing updated shortest paths} + O( N+2M) {for the adjacency list} ~ O(N+M) .

Where N= number of vertices and M= number of edges.